+++ sanely-jsoniter/src/sanely/jsoniter/SanelyJsoniter.scala
@@ -307,7 +307,57 @@ object SanelyJsoniter:
       else if tpe =:= TypeRepr.of[String] then Some(nullGuardWrite(efa))
       else if tpe =:= TypeRepr.of[BigDecimal] then Some(nullGuardWrite(efa))
       else if tpe =:= TypeRepr.of[BigInt] then Some(nullGuardWrite(efa))
-      else None
+      else
+        // Inline Option[prim] and List[prim] writes to avoid virtual dispatch
+        tpe match
+          case AppliedType(tycon, List(innerArg)) =>
+            def innerWriteVal(valTerm: Term): Option[Term] =
+              val ei = innerArg.dealias
+              if ei =:= TypeRepr.of[Int] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[Long] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[Float] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[Double] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[Boolean] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[Short] then Some(writeVal(Select.unique(valTerm, "toInt")))
+              else if ei =:= TypeRepr.of[Byte] then Some(writeVal(Select.unique(valTerm, "toInt")))
+              else if ei =:= TypeRepr.of[Char] then Some(writeVal(callMethod(valTerm, "toString", Nil)))
+              else if ei =:= TypeRepr.of[String] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[BigDecimal] then Some(writeVal(valTerm))
+              else if ei =:= TypeRepr.of[BigInt] then Some(writeVal(valTerm))
+              else None
+            tycon.typeSymbol.fullName match
+              case "scala.Option" =>
+                val owner = Symbol.spliceOwner
+                val optSym = Symbol.newVal(owner, "_o", tpe, Flags.EmptyFlags, Symbol.noSymbol)
+                val optDef = ValDef(optSym, Some(efa))
+                val optRef = Ref(optSym)
+                val isDefined = Select.unique(optRef, "isDefined")
+                val get = Select.unique(optRef, "get")
+                innerWriteVal(get).map { writeTerm =>
+                  Block(List(optDef), If(isDefined, writeTerm, writeNull()))
+                }
+              case s if s.endsWith(".List") =>
+                val owner = Symbol.spliceOwner
+                val lstSym = Symbol.newVal(owner, "_l", tpe, Flags.EmptyFlags, Symbol.noSymbol)
+                val lstDef = ValDef(lstSym, Some(efa))
+                val lstRef = Ref(lstSym)
+                val isNull = Apply(Select.unique(lstRef, "=="), List(Literal(NullConstant())))
+                val arrStart = callMethod(outTerm, "writeArrayStart", Nil)
+                val arrEnd = callMethod(outTerm, "writeArrayEnd", Nil)
+                val curSym = Symbol.newVal(owner, "_c", tpe, Flags.Mutable, Symbol.noSymbol)
+                val curDef = ValDef(curSym, Some(lstRef))
+                val curRef = Ref(curSym)
+                val nonEmpty = Select.unique(curRef, "nonEmpty")
+                val head = Select.unique(curRef, "head")
+                val tail = Select.unique(curRef, "tail")
+                innerWriteVal(head).map { writeHead =>
+                  val whileBody = Block(List(writeHead), Assign(curRef, tail))
+                  val whileLoop = While(nonEmpty, whileBody)
+                  val elseBranch = Block(List(arrStart, curDef, whileLoop), arrEnd)
+                  Block(List(lstDef), If(isNull, writeNull(), elseBranch))
+                }
+              case _ => None
+          case _ => None
 
     // === Sum derivation ===